+++ app/(nav)/todos/[todoId]/_view/NoteFormContent.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { ModalClose, ModalContent, ModalProvider, ModalTrigger } from '@/components/common/Modal';
 import TiptapEditor from '@/components/TiptapEditor';
 import useNoteMutation from '@/lib/hooks/useNoteMutation';
 import IconClose from '@/public/icons/IconClose';
@@ -31,13 +30,12 @@ import IconTextItalics from '@/public/icons/IconTextItalics';
 import IconTextNumberPoint from '@/public/icons/IconTextNumberPoint';
 import IconTextUnderline from '@/public/icons/IconTextUnderline';
 import IconTextHighlight from '@/public/icons/IconTextHighlight';
-import IconAddLink from '@/public/icons/IconAddLink';
-import InputSlid from '@/components/common/InputSlid';
 import Button from '@/components/common/ButtonSlid';
-import ModalSavedNote from './ModalSavedNote';
+import OpenSavedNoteModal from './OpenSavedNoteModal';
 import { afterNoteMutation } from '@/app/actions';
 import { useQueryClient } from '@tanstack/react-query';
 import SecondsTimer from './SecondsTimer';
+import AddLinkModal from './AddLinkModal';
 
 export type SavedNote = {
   title: string;
@@ -82,7 +80,6 @@ const NoteFormContent = memo(
     const todoTitle = searchParams.get('todo');
     const goalTitle = searchParams.get('goal');
     const { editor } = useCurrentEditor();
-    const [linkUrlValue, setLinkUrlValue] = useState(linkUrl);
     const [title, setTitle] = useState(initTitle);
     const titleRef = useRef<HTMLInputElement>(null);
     const { mutate } = useNoteMutation(todoId as string);
@@ -93,9 +90,7 @@ const NoteFormContent = memo(
       return note.savedAt;
     }, [savedNote, todoId]);
 
-    const handleChangeLinkUrlValue: ChangeEventHandler<HTMLInputElement | HTMLSelectElement> = (e) =>
-      setLinkUrlValue(e.target.value);
-    const handleSaveLinkUrl = () => onSaveLinkUrl(linkUrlValue);
+    const handleSaveLinkUrl = (linkUrlValue: string) => onSaveLinkUrl(linkUrlValue);
     const handleChangeTitle: ChangeEventHandler<HTMLInputElement> = (e) => {
       setTitle(e.target.value.length > 30 ? e.target.value.slice(0, 30) : e.target.value);
     };
@@ -208,7 +203,7 @@ const NoteFormContent = memo(
               <IconClose circleFill='fill-blue-500' className='cursor-pointer' />
             </button>
             <p className='font-semibold text-sm grow'>임시 저장된 노트가 있어요. 저장된 노트를 불러오시겠어요?</p>
-            <ModalSavedNote onOpenSaved={onOpenSaved} savedNote={savedNote} />
+            <OpenSavedNoteModal onOpenSaved={onOpenSaved} savedNote={savedNote} />
           </div>
         )}
         <hr />
@@ -285,30 +280,7 @@ const NoteFormContent = memo(
               </label>
             </div>
             <div className='grow flex justify-end'>
-              <ModalProvider>
-                <ModalTrigger type='button'>
-                  <IconAddLink className='cursor-pointer hover:bg-slate-100' />
-                </ModalTrigger>
-                <ModalContent className='w-full max-w-[520px] flex flex-col'>
-                  <div className='flex justify-between mb-6'>
-                    <h1 className='text-lg font-bold'>링크 업로드</h1>
-                    <ModalClose />
-                  </div>
-                  <InputSlid
-                    label='링크'
-                    type='text'
-                    placeholder='영상이나 글, 파일의 링크를 넣어주세요'
-                    className='mb-10'
-                    value={linkUrlValue}
-                    onChange={handleChangeLinkUrlValue}
-                  />
-                  <ModalClose asChild>
-                    <Button className='w-full' onClick={handleSaveLinkUrl}>
-                      확인
-                    </Button>
-                  </ModalClose>
-                </ModalContent>
-              </ModalProvider>
+              <AddLinkModal linkUrl={linkUrl} onSave={handleSaveLinkUrl} />
             </div>
           </div>
           {savedToast && (
